requirements.txt
colorama
pyvirtualcam
pyttsx3
colorama
cvzone
mediapipe
sounddevice
python_dotenv
datetime
opencv-python
face_recognition
discord_webhook
numpy
